OBP - H1_H2/requirements-3tier.txt
pyodbc
python-dotenv
pandas
fastapi
uvicorn
requests